components_repo/llm-as-a-judge-evals/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.8.0
openai>=1.0.0
anthropic>=0.34.0
google-genai>=0.5.0
